using the correct types

+++ src/contract.rs
@@ -73,14 +73,14 @@ fn execute_supply(deps: DepsMut, info: MessageInfo) -> Result<Response<OsmosisMs
         .add_attribute("available_funds", pool.available))
 }
 
-fn execute_swap(_deps: DepsMut, _info: MessageInfo, input: i32, min_output: i32) -> Result<Response<OsmosisMsg>, ContractError> {
+fn execute_swap(_deps: DepsMut, _info: MessageInfo, input: u128, min_output: u128) -> Result<Response<OsmosisMsg>, ContractError> {
     let swap = OsmosisMsg::simple_swap(
         1,
         "uosmo",
         "uion",
         SwapAmountWithLimit::ExactIn {
-            input: Uint128::from(input as u128),
-            min_output: Uint128::from(min_output as u128)
+            input: Uint128::from(input),
+            min_output: Uint128::from(min_output)
         }
     );
     let msgs = vec![SubMsg::new(swap)];
